pyproject.toml
[tool.poetry]
name = "eda-ja"
version = "0.1.0"
description = ""
authors = ["Yuji Kamiya <y.kamiya0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
nltk = "^3.6.5"
mecab-python3 = "^1.0.4"
pandas = "^1.3.4"
requests = "^2.26.0"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
